Add supported columns, add alphabetization of authorities. (#277)
Added two more supported columns from incentive-a-thon spreadsheet, and
added alphabetization of authorities' file. Typical json sorting wasn't
successful as the key's name was the value to sort by, which wouldn't
parse when defining the sort function.

That being said, a generalized alphabetization solution for both
programs and authorities is also being experimented with, but this
simpler solution is viable due to incentive calculation being sorted by
state locations of incentives.

Programs still need to be sorted, and README details this.