Improve spreadsheet-to-json value standardization. (#282)
This does a few things better than the current approach:

1. It switches VALUE_MAPPINGS to behave more like FIELD_MAPPINGS in
several respects. It's now passed in instead of being a global and it's
closer to the same format (canonical -> list of aliases).
2. Uses that easier-to-work-with format to rename values as part of
standardizing field names. I'll rename that method in a follow-up so as
to not touch even more files. As a result, all of the gross enum stuff
we had to do can be removed. I think the bracket syntax is temporary; I
have a follow-up that will make this all even cleaner.
3. Renames a bunch of columns in `spreadsheet-mappings.ts`. These were
done to be closer to the JSON schema (authority_level -> authority_type)
or to use a nested syntax like `amount.type` which will make from CSV
format to JSON a breeze.


Testing: added unit tests, plus I ran CO through this and I get the
exact same values as what's currently in the JSON.