Standardize capitalization for aliases. (#285)
I went with a "title case" approach, so this capitalizes most things but
not short articles or conjunctions. Happy to go with a more consistent
approach if you prefer.

A few other small tweaks, like switching the order of some aliases and
correcting a misspelling of weedwacker.

Regarding splitting up cases that contain different concepts: it's a bit
more complicated than I'd figured, in part because we might have
incentives live in prod that have been collapsed in a single concept
like `heat_pump_air_conditioner_heater` for ductless heat pumps. This is
fine for the time being, but we'll have to address it before we start
relying on the back-to-csv pathway.