Add schema for collected data and base our state incentives schema on that. (#284)
Main thing: adds schemas for collected data and derived data, and then
defines downstream schemas like the main StateIncentive in relation to
them. We don't exploit that in this PR, but it opens up the door for the
features we actually care about, like easier spreadsheet-to-json
conversion and tests that assess whether the spreadsheet data itself
meets our standards/has problems.

Side thing: I noticed that the amount schema is defined once for state
incentives and once for IRA incentives. i think it's best to consolidate
until we have a reason to split them.