Add format-converter to bridge from csv-parse to our collected schema. (#292)
This is not integrated yet – that is an upcoming PR.

I was tempted to call these methods csvToJson, but really csv-parse has
already converted csv to json; it's just a flat object that we can't do
interesting things with. These methods "inflate" that object and
optionally validate it.

I noticed an error or two in the previous schema as testing, so those
are fixed here.