Add testfilter yarn script for running a subset of tests. (#295)
Our test suite's grown a lot, and it would be nice to have an easy way
to only run some tests while working on individual files. `yarn
testfilter <path>` does this.

In yarn 2, there's a slicker way to do this that doesn't require a
separate command, but I couldn't figure out an easy way to do this in
yarn 1.