Validate spreadsheets first in spreadsheet-to-json flow. (#293)
This incorporates the format-converter and collected schema into the
spreadsheet-to-json flow. Now we first validate that the collected data
is valid under our spreadsheet schema, then take the collected data and
turn it into a refined StateIncentive. This may surface some more nits
in the short-term while we clean up the spreadsheets, but in the
long-term is better since we can keep those spreadsheets healthy from
the get-go.

It simplifies the `refineCollectedData` method since we can assumed
we're working with valid collected data from the beginning. However, it
did require rewriting a bit of the LowIncomeThreshold part since we
could no longer just assign a string.

After this PR, I'll also pull out a stand-alone version of the
CollectedFields validation that can be run on our spreadsheets on a
periodic basis so we can be aware of any spreadsheet validity
regressions.