Integrate ENERGY STAR® utility dataset (#286)
## Description

This pulls in the mapping from zip codes to utilities in the dataset
here:

https://www.energystar.gov/buildings/owners_and_managers/existing_buildings/use_portfolio_manager/find_utilities_provide_data_benchmarking

The centerpiece here is the script that converts that Excel file into
two CSV files that are then imported into the API server's SQLite DB
for runtime querying. It takes the human-friendly utility names from
the sheet, does a bit of cleaning, then turns them into utility IDs in
the same style as we have in the code already
(e.g. `ri-rhode-island-energy`).

The script is not run at build time. It's meant to be run by hand, and
the resulting CSVs checked in, like everything else in `scripts/data`.

For now, the behavior is that the result set is the _intersection_ of
the dataset's result and our own `authorities.json`. This is a
conservative approach, avoiding returning any utility we aren't
already returning today. This may not be the long-term final state; I
think eventually we may not want utilities in `authorities.json` at
all, and instead consider this dataset as canonical.

The user-visible names are still from `authorities.json`, not from the
dataset.

This does mean that we will _stop_ returning some "utilities" that
we're returning today, albeit only in beta states. These are
consortiums of small utilities (e.g. VPPSA in VT, Tri-State G&T in CO)
that offer incentives to customers of their member utilities.
Resolving this problem will entail separating the concept of "who
offers an incentive" from "which utility's customers are eligible".

While going through the dataset's utilities for the states we have
incentive data for, I made some changes to our names/IDs, to better
reflect the way the utilities are branded to customers. The changes in
`data` are all straight find/replace.

## Test Plan

`yarn test`. This includes a new test that ensures all of our IDs in
authorities.json exist in the dataset (except for the exceptions, as
noted above).

Spot-checking with ZIP codes in Rhode Island (where we know the
situation pretty well) and other beta states, especially focusing on
zips around towns where there are municipal utilities.