Add spreadsheet health test and register CO and AZ as passing states. (#301)
This test validates that the spreadsheets match our collected data
schema. It is skipped right now, but likely next step is to add a
workflow that sets an environment variable so we can run it only in
those contexts.

CO and AZ are the only tests currently passing :( but I will circle back
to the others later.