Adding Spanish descriptions for CT (#302)
I also increased the character limit for Spanish descriptions in the
schema test from 160 to 200. On Slack, we had discussed increasing to
180 based on the idea that on average Spanish is 20% longer than
English. However, even 200 still doesn't cover a few of the longer
descriptions. I left those out for now but they should be added back
when we have a better way to display longer descriptions.


[Here's](https://docs.google.com/spreadsheets/d/13mJyltzLt4KTDuuL14r71HqKJU5c9f6bYJtGLotgkVI/edit?pli=1#gid=995688950)
the spreadsheet with the descriptions for reference.