Document generate-utility-data script in README (#303)
## Description

I realized I was going through a somewhat finicky process to clean up
the utility data for each state, and this is now has to be an integral
part of the spreadsheet-to-JSON process, so I'd better document it.

I'm not yet totally sure where this process should fit into the
overall process for each state, so interested in feedback on that.

Also I realized I have no plan for running this script on a regular
basis. I have no idea how often ENERGY STAR updates the data, so I
thought the script should at least print out the last-updated date (in
the first cell of the spreadsheet) so we can start to figure that out.

## Test Plan

Read it. Run the script.