Refactor spreadsheet-to-json to avoid writing files. (#306)
We get better reuse of the spreadsheet-to-json process if decoupled from
the file reading/writing parts. So this just moves the data processing
into a single method and then file reading/writing happens separately.
As part of that, we don't error out early when we find
CollectedIncentive errors; we continue to try to validate
StateIncentives and surface both separately to the user if they exist.

I also realized the updateJsonFiles method handles file-clearing already
so a bit of tidying up there too.