[RFC] Add "utilities" list field to incentives
I've been messing around with this for days, and it's time to just get
something out there and see what people think. This just does the
modification to the PRPA incentives in Colorado, since it was the
simplest and I'm not fully convinced by this approach.

I find this awkward in a few ways:

- I can't picture how the mapping from spreadsheet to JSON would work
  here. For spreadsheet purposes it seems most straightforward just to
  have "PRPA", but where in the spreadsheet-to-JSON process would this
  new field be introduced?

- It's duplicative in data; every incentive from one of these
  consortiums will have the same list of utilities. Not a huge
  problem, but still.

- It's duplicative in functionality; there'd now be two ways to
  specify an incentive that's offered to customers of a single
  utility.

- It doesn't cleanly extend to groups of counties (a thing that exists
  in NV, for example). Sure you could add an analogous `counties`
  field, but what would it mean for an incentive to have both a
  `counties` and a `utilities` field?

### Alternatives?

Alternative approaches to this all fall into the category of "define
authority groups somehow". That solves the duplicative-data problem,
but introduces all sorts of awkwardness in deciding how to represent
the groups.

- Is a group of authorities itself a kind of authority, or a distinct
  concept?

- Could a group include different kinds of authority, like some
  utilities and some counties? What would that even mean -- "one of
  these utilities AND one of these counties", or "one of these
  utilities OR one of these counties"?

I'm starting to form a thought on this that what's making things
difficult is the overlap of namespaces: utilities, cities, and
counties are each a namespace of their own, and an authority is its
own identity (name, logo) that can consist of one or more of the
above. That's just an awkward thing to represent.