Az update (#318)
Like
https://github.com/rewiringamerica/api.rewiringamerica.org/pull/316, I'm
doing this separate from the test to not muddy up the diffs. Having the
spreadsheet in line with JSON is good in any case though.

I mostly updated the spreadsheet, which had changes that were made on
the way to JSON that didn't get propagated back up.

The changes here are:
- pulling in Spanish descriptions (cc @mariabajzek)
- renaming programs and authorities to be the ones we would compute now
automated-ly
- field reordering
- in the spreadsheet, Salt River had two nearly identical program names,
one with (EV) and one without. While I was fixing it, I thought the name
without EV looked better (and is also what we'd use to get the shorter
program identifier)
- adding a start_date that didn't get picked up before