Update CO data to align with spreadsheet. (#316)
In preparation of adding stricter tests that verify spreadsheet and JSON
data are aligned, I'm updating the CO data. I thought about doing it in
the same PR, but it seemed like it would muddy the diffs a bit.

This was the result of me changing the spreadsheet data and then
rerunning the spreadsheet-to-json flow. Other states may have more
varied changes, but here we just have:

- authority and program renames
- field reordering

I didn't see any actual content changes aside from these.