Add Spanish descriptions for VA (#319)
This PR also:

- updates the update-descriptions script to use the spreadsheet
standardizer, since the VA spreadsheet column headers had newlines in
them so the old script couldn't read them. Let me know if I didn't use
the standardizer in an appropriate way, but this seemed like the
simplest update.
- updates the test fixtures, which in this case only included the CO
file.
- alphabetizes the states in the spreadsheet registry, to make it easier
to read as it grows.

**Test Plan**
Updated the test fixtures and made sure the unit tests still pass.