Move spreadsheet health test. (#326)
I saw that
https://github.com/rewiringamerica/api.rewiringamerica.org/pull/315 is a
better approach for having a test that is run as a one-off vs the
current approach of conditioning on an external boolean, so I'm adopting
that here. I'll add a GitHub action for it later.

Unrelated:
https://github.com/rewiringamerica/api.rewiringamerica.org/pull/319
accidentally turned the test off for AZ and on for RI (which fails), so
I'm reverting that.