Adding Spanish descriptions for VT (#333)
I ran the update-descriptions script to do this, but as a result I had
to revert many of the English descriptions because they are too long in
the spreadsheet and don't currently match what is in the JSON. Before
launching this, we should probably review the descriptions and decide
whether we want to keep the longer versions (and update the UI to
display them better) or keep the shorter versions and make sure the
spreadsheet is back in sync.

I also had to increase the Spanish character limit in the schema test
again because many of the Spanish ones are very long for VT.

Also, the spreadsheet <--> JSON translation is going to be very
complicated because of the EV incentives (e.g. VT-27). They are
represented as a single spreadsheet row with one description that maps
to multiple JSON objects with different amounts and descriptions based
on tax filing status. That will need to be fixed before this is easily
maintainable going forward.

**Test plan:**
Updated the test fixture to match the new data.