Add new update programs method now that programs.json is consolidated. (#338)
This is admittedly a bunch of string-hacking, but I don't see a better
way to do it, aside from giving up on this step being automated – which
is not out of the question since it's not run as often as the actual
incentive-spreadsheet-to-json script. Still, I think we might as well
give it a shot.

Testing: added a bunch of unit tests, and ran on CO and AZ to ensure
that we saw no diffs.