Add sync test that confirms that spreadsheet and JSON data is aligned. (#317)
This is more-or-less what I demo'd on Monday. We're pretty close to a
solution of this shape here, so putting out this PR to spur discussion
on some of the finer points – or to surface major philosophical
objections. The main files to review are the test itself and the readme;
the other changes are a bit ancillary.

This test – which replaces a weaker version that verified all data
matched our collected schema – validates that all rows in a spreadsheet
are in sync with checked-in JSON. It does so by basically running the
spreadsheet-to-json flow and verifying that the output is the same as
what's checked in. That means we will start checking in any records that
fail our schema validations, though I think that's a good thing as it's
a way of documenting gaps in our process.

This can't be a blocking test since it can start to fail due to external
factors (namely someone updating the spreadsheet without changing the
JSON!), and we don't want to halt other PRs because of that. My gut
reaction is that we can start with it being run as needed when updating
a state. Eventually, we can add an automatic daily (or more frequent)
run – this won't stop anyone's workflow, but it'll give the team a
heads-up that something needs to be attended to.

We can turn this on gradually, state by state, only for the states where
maintaining the test is easy. I suspect that only CO passes right now,
but CT and AZ are close.