Turn start/end dates into strings
## Description

This is mechanical: just stringifying the `start_date` and `end_date`
of every incentive record, and changing the v1 API to serve them as
strings instead of numbers.

In the process, I made the slightly horrifying discovery that `t.same`
in Tap considers values to be "same" up to type coercion; that is,
`2024` and `"2024"` are considered "same". I changed usages of that
test method to `strictSame`, which doesn't do type coercion.

I'll follow up with a PR to be smarter about parsing the values that
are in spreadsheets -- some state sheets have precise dates.

The new embed frontend has already been updated to handle the dates
being either numbers or strings. (It currently doesn't even read them,
and in fact the server code is hard-overwriting all state incentives'
start/end dates with 2023 and 2024 respectively. I'll be changing the
latter shortly.)

## Test Plan

`yarn lint` and `yarn test`. New test to validate the regex I'm using
to enforce date format in the JSON files.