Switch to configuring low-income explicitly by record. (#350)
This isn't fully complete, but I wanted to get initial feedback before
writing tests.

After doing some investigation, I didn't want to go the geo_groups route
for low_income configuration for a couple reasons:
1. low-income thresholds are of interest to external stakeholders, and I
think it will complicate things for all parties if at some point in the
spreadsheet-to-json process, we're replacing data that is standalone and
human-readable with an enum key that only makes sense if referencing a
separate dataset (this reasoning applies to geo groups too but is maybe
less important there)
2. in the spreadsheets, data collectors tend to put notes in the income
column that aren't shared across all records that have the same
low-income thresholds. Some of these refer to the non-low-income
counterpart incentive, if one exists (see CT and RI for examples). There
may be other use cases too. I don't want to collapse these notes into
the same JSON object like we did with geo_groups.

What I favor instead is keeping low-income in spreadsheet form as a
free-text field, and then using config to tie the thresholds to records.
I chose records instead of authorities or programs because I don't think
it's that much extra config and it's more precise.

While it's redundant to express the records in the thresholds and also
have low_income as an incentive field, I think it makes sense since
creation of the incentives JSON is automated. In that sense we still
have a single representation of low-income information and then the
incentives JSON is downstream of that.

If this is okay, I'll add static tests to verify that the info in the
thresholds and JSON reflect the same reality.

I'm also getting rid of the default requirement since we won't need that
in the future.