Try to parse full dates from spreadsheets; make dates optional in v1 (#356)
## Description

We'll want to change the spreadsheets' guidance for how to input
dates, but for now, it's easy enough to parse all the date formats
that people have input.

I chose this datetime library because it's designed to resemble
`java.time`, which is the most sensible datetime library I've used.

Separately, it's now pretty clear to me that it doesn't make sense to
have the date fields be required in the v1 API -- most state
incentives don't have dates defined, and so far we've just been
hardcoding in `start_date: 2023` and `end_date: 2024` even for
incentives that do define their own dates. (Fortunately the frontend
isn't even reading them.) So they're now optional.

## Test Plan

`yarn lint` and `yarn test`.

I ran the spreadsheet-to-JSON script on AZ, CO, and VT, which are the
only ones that have any incentives with dates in them, and made sure
there were no failures due to unparseable dates. AZ and CO succeeded;
changes are reflected here. VT failed for unrelated reasons; will look
into that separately.

Added an unparseable date to an incentive in the CO sheet and made
sure the script output that row as invalid.