v0: add new more_info_url_internal field to expose new federal IRA incentive pages (#361)
Asana ticket:
https://app.asana.com/0/1206540452900052/1206673169101907/f

## What

The consumer app team is releasing a bunch of new pages with information
on the IRA incentives. We want to update the federal IRA incentive
"learn more" links to point to these new pages.

Note that none of these pages are live yet and won't be until Monday.

## Releasing

This PR is safe to be released today because it's only modifying the
data in `ira_incentives.json` and adding a new field to v0 to expose
these links. It shouldn't make any changes to v1 or to the existing v0
clients.