Make the Energy Star utility data authoritative
## Description

Use the Energy Star dataset to generate authorities.json, rather than
having authorities.json hold a hand-curated set of utilities. We can
still manually add logos to utilities in authorities.json, but their
names will come from Energy Star.

The zip-to-utilities table in SQLite is unchanged, but the `utilities`
table is now gone. This changes the initial curation workflow a little
bit -- slightly for the worse in my opinion -- but overall I think
it's still fine. We'll only have to do it once per state.

This will result in us showing a bunch of utilities for which we
haven't done data collection at all, or which we've confirmed do not
offer any incentives. The justification for doing this is that we'll
be much more likely to let users choose their actual utility, rather
than having to pick the wrong answer, or an uninformative "Other"
option. If we decide that we want to be able to distinguish between
these kinds of utilities (still an open product question), we can
easily add a field to the struct in authorities.json to capture those
distinctions.

## Test Plan

`yarn lint` and `yarn test`.

Look at the diff to authorities.json. (There's some churn because I'm
sorting the utilities by ID now, but there shouldn't be such churn in
future.)

Query `/api/v1/utilities` for a sample zip code that was previously
not returning any utility results (89040) and make sure there's a
result now, and that querying `/calculator` with that utility ID
doesn't error. (There won't be any incentives for that utility, but
the point is the API accepts it as a valid utility ID.)