Remove location parameter from v1
## Description

This is the final step of migrating to separate top-level `zip` and
`address` parameters: removing the old nested parameter.

The embed frontend is already switched over, and there should be no
other clients of v1, so although this is a breaking change it should
be safe.

## Test Plan

`yarn lint`, `yarn test`. Run the update-fixtures script and make sure
the updates all succeed and there are no changes.