Add Pennsylvania incentives (#371)
Adds PA incentives. Spreadsheet:
https://docs.google.com/spreadsheets/d/1H4hVUAwmEshbndeY2lFdnyk53qPZhoR5o1du5G1DVMw/edit#gid=616584455

Note this spreadsheet was initially generated by the LLM scripts but I
manually reviewed and updated every incentive. Because this data is all
new, it should all be up to date and ready to be released/tested by
external partners as soon as this is in.

Ticket: https://app.asana.com/0/1206629812394927/1206778652392945/f