Make <p> tag the main content for hashing in HTML.
I've gone through a decent amount of incentive URLs, and have found for non-PDFs, links with HTML have important info contained in <p> tags that will likely be enough to tell when content changes.

Exhaustive testing is probably a good step forward for diffing these, but this helps get it to a general state with a couple false negatives rather than numerous false positives.