Add way to read Google Sheets API and save hyperlinks into JSON. (#373)
This is half of the Google Sheets <-> JSON conversion. The other half is
coming shortly.

Welcoming @mmezher's thoughts but @oyamauchi, I think it's best if
you're the primary reviewer. Since you'll be absorbing any
spreadsheet-json work, I want to make sure you're on board.

As mentioned in [this
thread](https://rewiringameri-g3x1100.slack.com/archives/C04E4THEGMT/p1710441552470399),
hyperlinks are a problem for the conversion since they are not easily
stored in JSON. This is especially true because you can have multiple
links in the same cell, and it becomes difficult to store this without
making it very difficult for humans to edit the JSON. So as decided in
that thread, we're condensing the hyperlinks to text.

To actually deploy this new method, I've changed the spreadsheet-to-json
flow so that we start saving the CollectedIncentives (spreadsheet
equivalents) for states that specify an additional collectedFilepath in
the spreadsheet registry. There's also some Google API setup. I haven't
documented this much in the README; I'm hoping to get the second half in
place before making this widely available.