Split authorities.json into a file for each state (#385)
## Description

We're about to populate authorities for every state's utilities in one
go, and that would make authorities.json unmanageably large. Since its
top-level key is already states, that makes it easy to split into one
file per state.

There's basically no downstream effect in the API server;
`authorities.ts` just recreates the same map that was previously in
the unified authorities.json. The scripts side is actually simplified
a bit.

The `state` key in a state's authorities is now optional, because we'll
be generating files with only a `utility` key.

https://app.asana.com/0/1206661332626418/1206882125337116

## Test Plan

`yarn lint` and `yarn test`.

Followed the new-state process (running generate-utility-data and then
authority-and-program-updater) for CA to make sure it goes smoothly.