Add method to export collected incentives to a Google sheet. (#374)
This is the other half of Google Sheet <-> JSON conversion; a
continuation of
https://github.com/rewiringamerica/api.rewiringamerica.org/pull/373.
They're largely independent, though they both rely on the authentication
helper and the changes to the spreadsheet-registry.

This is the CollectedIncentives -> Google Sheet direction. It uses the
Google Sheets API to create a new spreadsheet. Right now, we only have
the incentives data itself, but as a follow-up, I'll create other
worksheets like the data model and the standardized enum values so that
it is more comparable to our existing sheets (plus add data validation).

Notable assumption: the first alias in the FIELD_MAPPINGS and
VALUE_MAPPINGS in spreadsheet-mappings.ts is considered the canonical
human-readable spreadsheet version. I think this is fine for now, though
we may want clearer documentation of it at some point.

Testing: unit tests, and I tried this on Colorado. You can see the
round-trip in [this
spreadsheet](https://docs.google.com/spreadsheets/d/1vybxSsCChFp1iAwMy1CwUFEjAxh0KxAmoDl7grtlKQ8/edit#gid=406119121).
This was running the Sheets -> CollectedIncentives version followed by
CollectedIncentives -> Sheets, then copying over the original CO data
and comparing. The green cells are where there's a difference (the new
value is on top).

 The cell values are the same, except in:
1. places where the original Colorado data used a different alias from
the canonical alias
2. dates. The original sheet has these as numbers formatted as dates;
the new version has them as strings. In keeping with Owen's direction
for dates, I think we want them to be strings anyway.
3. columns missing from the Colorado spreadsheet - there are some, like
Editorial Notes, that aren't in the original CO data, but in the
interest of standardizing our spreadsheets, I think it's good to be
adding them back in.
4. Cells that had hyperlinks (as expected)