Move project initialization into a function and cut test time in half. (#399)
I noticed recently that test time had ballooned – I was getting like 45
seconds to run all the tests. Did some debugging and the culprit was
initializing the TS project at the top-level of this file instead of
moving it into a function and doing as needed. Now we're back down to 23
seconds.