MI data refresh updates (#393)
Data refresh for Michigan. Main changes:
- Combine a few rebates, for example heat pumps of different
efficiencies. It seems like that's the direction we're moving towards
- Update with 2024 programs since several incentives have changed or
expired

### Open questions
MI-1 through MI-18 are expired. I marked as expired in the spreadsheet
and the end date is end of 2023, but they still show up in the
calculator. I'm guessing fixing this is part of the date work, but do we
want to just delete these from the spreadsheet in the meantime?