Account for Colorado income tax in state tax credit eligibility/maximums (#394)
## Description

Colorado offers tax credits against state taxes. This PR does as we do
with federal taxes and calculates the user's tax burden to avoid showing
tax credit incentives if the user wouldn't be able to claim them because
they don't owe enough tax.

The state has a flat tax rate of 4.4% and does not have a standard
deduction.

[Asana](https://app.asana.com/0/1205271193902672/1206661332626430/f)

## Test plan

- added tests! `yarn test`