Remove item.url from v1 (#410)
## Description

This is no longer used by the frontend, and the URLs now get
redirected to incentive-centered pages; I think the whole concept of a
URL for an item is dead.

v0 still uses the URLs, in the `more_info_url` field, so we can't
delete them. We can't even change them to the incentive-centric
`homes.rewiring` URLs, because these URLs are relative, and the host
is assumed to be `www`. So we'll be stuck with this until v0 can be
deleted.

## Test Plan

`yarn test`. Make sure the embed frontend does not use this field.