Small refresh of CO data based on Zero Homes feedback. (#413)
There are a bunch of diffs here that are unrelated to the refresh
because the spreadsheet and JSON got out of sync :)

Refresh Diffs:
1. Two incentives added at the end – CO-543 and CO-544
2. Some incentives changed in the middle because their heat pump terms
changed from 2023 to 2024.
3. Introduced two new low-income incentives which are bonuses
4. Updates of related files (authorities, programs, and test fixtures)

Unrelated diffs:
1. Fixing some typo spaces at the end of sentences
2. Fixing some cases where the text said "see link", which were being
lost anyway (these were actual links and not just a reference to the
utility's website). I don't think any of them were necessary.
3. Intentional loss of an incentive that we deemed had too little
information to bother showing. Before we had it as "Other"; now we're
using the "Omit from API?" flag to actually drop it