Add utility data for all states
## Description

Since we want to show a utility selector for everybody, we need
utilities for all the states. So here they are!

I did not QA these as intensively as I did the initial set of
states. I figure whoever adds each subsequent state's incentives
should do that, as I documented in the README.

I didn't Google every single utility. I looked for weird acronyms or
artifacts, or utilities that obviously seemed to be the same one under
different names. I made some changes to the name-cleaning logic to fix
a lot of the common cases automatically (which is why there are
changes to the GA and PA files).

The California file, and new overrides/exclusions for CA, are taken
from #380.

### Followups

- Remove the `include_beta_states` conditioning from `/v1/utilities`
  to allow utility selection in all states. (That maybe should wait
  until after the frontend has the "Other" option added.)

- Add territories. The Energy Star data doesn't cover them. However,
  fortunately, the utility situation in the territories is not super
  complicated.

https://app.asana.com/0/1206661332626418/1206843501113809

## Test Plan

Run the script to produce this data. Make sure running it again
doesn't change anything. QA the resulting data as described above.