Add Pennsylvania DEP EV rebate incentives (#419)
Our Duquesne point of contact let us know about this program and asked
if we can add these state level EV rebates. These rebates have a couple
of permutations:
- They apply to new and used EVs (but slightly different criteria - for
example, a new EV can be leased and still be eligible, used has to be
purchased)
- They apply to battery EVs and plug in hybrids
- All of the incentives are income-qualified (400% of poverty level),
but there are additional incentives available for residents at 200% of
the poverty level.

I ended up adding 4 variations of this incentive to cover all of the
cases:
PA-47 - New EV or PHEV (income qualified, 400% of poverty level)
PA-48 - Used EV or PHEV (income qualified, 400% of poverty level)
PA-49 - New EV or PHEV with bonus incentives (income qualified, 200% of
poverty level). Users not eligible for PA-47 if they're eligible for
this one
PA-50 - Used EV or PHEV with bonus incentives (income qualified, 200% of
poverty level). Users not eligible for PA-48 if they're eligible for
this one