Update Vermont data
## Description

I spent a while getting the Vermont sheet ready for the sheet-to-JSON
process. (As one of our earliest sheets, it's never been through this
process before.) Mainly:

- Adding necessary columns

- Fixing program names and making program URLs match the names

- Cutting down long descriptions

- Making per-filing-status and per-income-bracket copies of VT-27

- Excluding some incentives because they're expired, too vague, or not
  something we currently represent (the latter being the "replace my
  ride" ICE-car-scrapping incentives)

- Tweaking authority types (we're saying EVT is a state-level
  authority for these purposes) and adding the geo-eligibility column

- Adding prereq relationships for income-eligible bonuses

I used `scripts/authority-and-program-updater.ts` and
`incentive-spreadsheet-to-json.ts` and this is the result.

I didn't do any source-material research in this pass; I was just
getting the sheet into shape. Will follow up by looking at specific
issues raised in this Slack thread:
https://rewiringameri-g3x1100.slack.com/archives/C06FLDFSDDF/p1713290760370369?thread_ts=1713192707.714229&cid=C06FLDFSDDF

## Test Plan

`yarn test`